tests for delete event table

+++ controller/event.controller.js
@@ -47,6 +47,15 @@ exports.updateEventById = async (req, res) => {
 
 exports.deleteEventById = async (req, res) => {
   const { id } = req.params;
-  const eventDeleted = await Event.destroy({ where: { id: id } });
-  res.status(200).json(eventDeleted);
+  try {
+    const findEventById = await Event.findByPk(id);
+    if (!findEventById) {
+      res.status(404).json({ error: 'Entry not found.' });
+    } else {
+      const eventDeleted = await Event.destroy({ where: { id: id } });
+      res.status(204).json(eventDeleted);
+    }
+  } catch (err) {
+    res.status(500).json(err);
+  }
 };
